+++ src/indicators/mod.rs
@@ -1,6 +1,7 @@
 mod atr;
 mod bb;
 mod ema;
+mod kc;
 mod macd;
 mod rsi;
 mod sma;
@@ -9,6 +10,7 @@ mod stoch;
 pub use atr::*;
 pub use bb::*;
 pub use ema::*;
+pub use kc::*;
 pub use macd::*;
 pub use rsi::*;
 pub use sma::*;